refactor: convert UserIcon to TypeScript
