+++ src/components/SharedComponents/InlineUser/InlineUser.js
@@ -6,7 +6,7 @@ import INatIcon from "components/SharedComponents/INatIcon";
 // Directly imported, not from index.js to avoid circular dependency
 import Body3 from "components/SharedComponents/Typography/Body3.tsx";
 // Directly imported, not from index.js to avoid circular dependency
-import UserIcon from "components/SharedComponents/UserIcon/UserIcon";
+import UserIcon from "components/SharedComponents/UserIcon.tsx";
 import {
   Pressable, View
 } from "components/styledComponents";
