+++ src/components/SharedComponents/index.js
@@ -79,6 +79,6 @@ export { default as Subheading2 } from "./Typography/Subheading2";
 export { default as UnderlinedLink } from "./Typography/UnderlinedLink";
 export { default as UploadProgressBar } from "./UploadProgressBar";
 export { default as UploadStatus } from "./UploadStatus/UploadStatus";
-export { default as UserIcon } from "./UserIcon/UserIcon";
+export { default as UserIcon } from "./UserIcon";
 export { default as UserText } from "./UserText";
 export { default as ViewWrapper } from "./ViewWrapper";
